Storn changes on BarcodeReader and appsettings changes

+++ DBModel/Models/ProfitLossReport.cs
@@ -12,10 +12,10 @@ namespace DBModel.Models
         int prosecanTrosakPoBeracu;
         int trosak;
         decimal netoOtkup;
-        int prosecnaPC;
+        decimal prosecnaPC;
         int prosecnaCenaKost;
-        int prihod;
-        int profit;
+        decimal prihod;
+        decimal profit;
         int trosakProc;
 
         public DateTime Datum { get => datum; set => datum = value; }
@@ -24,10 +24,10 @@ namespace DBModel.Models
         public int ProsecanTrosakPoBeracu { get => prosecanTrosakPoBeracu; set => prosecanTrosakPoBeracu = value; }
         public int Trosak { get => trosak; set => trosak = value; }
         public decimal NetoOtkup { get => netoOtkup; set => netoOtkup = value; }
-        public int ProsecnaPC { get => prosecnaPC; set => prosecnaPC = value; }
+        public decimal ProsecnaPC { get => prosecnaPC; set => prosecnaPC = value; }
         public int ProsecnaCenaKost { get => prosecnaCenaKost; set => prosecnaCenaKost = value; }
-        public int Prihod { get => prihod; set => prihod = value; }
-        public int Profit { get => profit; set => profit = value; }
+        public decimal Prihod { get => prihod; set => prihod = value; }
+        public decimal Profit { get => profit; set => profit = value; }
         public int TrosakProc { get => trosakProc; set => trosakProc = value; }
     }
 }
